+++ core/lib/Thelia/Action/BaseCachedFile.php
@@ -14,6 +14,7 @@ namespace Thelia\Action;
 use Thelia\Core\Event\CachedFileEvent;
 use Thelia\Core\Event\File\FileCreateOrUpdateEvent;
 use Thelia\Core\Event\File\FileDeleteEvent;
+use Thelia\Core\Event\File\FileToggleVisibilityEvent;
 use Thelia\Core\Event\UpdateFilePositionEvent;
 use Thelia\Exception\FileException;
 use Thelia\Files\FileManager;
@@ -250,6 +251,11 @@ abstract class BaseCachedFile extends BaseAction
     public function updatePosition(UpdateFilePositionEvent $event)
     {
         $this->genericUpdatePosition($event->getQuery(), $event);
+    }
+
+    public function toggleVisibility(FileToggleVisibilityEvent $event)
+    {
+        $this->genericToggleVisibility($event->getQuery(), $event);
     }
 
 }