+++ core/lib/Thelia/Core/Template/Loop/Document.php
@@ -17,6 +17,7 @@ use Thelia\Core\Template\Loop\Argument\Argument;
 use Thelia\Core\Event\Document\DocumentEvent;
 use Thelia\Core\Event\TheliaEvents;
 use Thelia\Core\Template\Loop\Argument\ArgumentCollection;
+use Thelia\Type\BooleanOrBothType;
 use Thelia\Type\TypeCollection;
 use Thelia\Type\EnumListType;
 use Propel\Runtime\ActiveQuery\Criteria;
@@ -51,6 +52,7 @@ class Document extends BaseI18nLoop implements PropelSearchLoopInterface
 
                 Argument::createIntListTypeArgument('id'),
                 Argument::createIntListTypeArgument('exclude'),
+                Argument::createBooleanOrBothTypeArgument('visible', 1),
                 new Argument(
                         'order',
                         new TypeCollection(
@@ -199,8 +201,14 @@ class Document extends BaseI18nLoop implements PropelSearchLoopInterface
         }
 
         $exclude = $this->getExclude();
-        if (!is_null($exclude))
+        if (!is_null($exclude)) {
             $search->filterById($exclude, Criteria::NOT_IN);
+        }
+
+        $visible = $this->getVisible();
+        if ($visible !== BooleanOrBothType::ANY) {
+            $search->filterByVisible($visible ? 1 : 0);
+        }
 
         return $search;
 
@@ -239,6 +247,7 @@ class Document extends BaseI18nLoop implements PropelSearchLoopInterface
                     ->set("CHAPO"                 , $result->getVirtualColumn('i18n_CHAPO'))
                     ->set("DESCRIPTION"           , $result->getVirtualColumn('i18n_DESCRIPTION'))
                     ->set("POSTSCRIPTUM"          , $result->getVirtualColumn('i18n_POSTSCRIPTUM'))
+                    ->set("VISIBLE"               , $result->getVisible())
                     ->set("POSITION"              , $result->getPosition())
                     ->set("OBJECT_TYPE"           , $this->objectType)
                     ->set("OBJECT_ID"             , $this->objectId)