+++ core/lib/Thelia/Model/Base/BrandDocumentQuery.php
@@ -25,6 +25,7 @@ use Thelia\Model\Map\BrandDocumentTableMap;
  * @method     ChildBrandDocumentQuery orderById($order = Criteria::ASC) Order by the id column
  * @method     ChildBrandDocumentQuery orderByBrandId($order = Criteria::ASC) Order by the brand_id column
  * @method     ChildBrandDocumentQuery orderByFile($order = Criteria::ASC) Order by the file column
+ * @method     ChildBrandDocumentQuery orderByVisible($order = Criteria::ASC) Order by the visible column
  * @method     ChildBrandDocumentQuery orderByPosition($order = Criteria::ASC) Order by the position column
  * @method     ChildBrandDocumentQuery orderByCreatedAt($order = Criteria::ASC) Order by the created_at column
  * @method     ChildBrandDocumentQuery orderByUpdatedAt($order = Criteria::ASC) Order by the updated_at column
@@ -32,6 +33,7 @@ use Thelia\Model\Map\BrandDocumentTableMap;
  * @method     ChildBrandDocumentQuery groupById() Group by the id column
  * @method     ChildBrandDocumentQuery groupByBrandId() Group by the brand_id column
  * @method     ChildBrandDocumentQuery groupByFile() Group by the file column
+ * @method     ChildBrandDocumentQuery groupByVisible() Group by the visible column
  * @method     ChildBrandDocumentQuery groupByPosition() Group by the position column
  * @method     ChildBrandDocumentQuery groupByCreatedAt() Group by the created_at column
  * @method     ChildBrandDocumentQuery groupByUpdatedAt() Group by the updated_at column
@@ -54,6 +56,7 @@ use Thelia\Model\Map\BrandDocumentTableMap;
  * @method     ChildBrandDocument findOneById(int $id) Return the first ChildBrandDocument filtered by the id column
  * @method     ChildBrandDocument findOneByBrandId(int $brand_id) Return the first ChildBrandDocument filtered by the brand_id column
  * @method     ChildBrandDocument findOneByFile(string $file) Return the first ChildBrandDocument filtered by the file column
+ * @method     ChildBrandDocument findOneByVisible(int $visible) Return the first ChildBrandDocument filtered by the visible column
  * @method     ChildBrandDocument findOneByPosition(int $position) Return the first ChildBrandDocument filtered by the position column
  * @method     ChildBrandDocument findOneByCreatedAt(string $created_at) Return the first ChildBrandDocument filtered by the created_at column
  * @method     ChildBrandDocument findOneByUpdatedAt(string $updated_at) Return the first ChildBrandDocument filtered by the updated_at column
@@ -61,6 +64,7 @@ use Thelia\Model\Map\BrandDocumentTableMap;
  * @method     array findById(int $id) Return ChildBrandDocument objects filtered by the id column
  * @method     array findByBrandId(int $brand_id) Return ChildBrandDocument objects filtered by the brand_id column
  * @method     array findByFile(string $file) Return ChildBrandDocument objects filtered by the file column
+ * @method     array findByVisible(int $visible) Return ChildBrandDocument objects filtered by the visible column
  * @method     array findByPosition(int $position) Return ChildBrandDocument objects filtered by the position column
  * @method     array findByCreatedAt(string $created_at) Return ChildBrandDocument objects filtered by the created_at column
  * @method     array findByUpdatedAt(string $updated_at) Return ChildBrandDocument objects filtered by the updated_at column
@@ -152,7 +156,7 @@ abstract class BrandDocumentQuery extends ModelCriteria
      */
     protected function findPkSimple($key, $con)
     {
-        $sql = 'SELECT `ID`, `BRAND_ID`, `FILE`, `POSITION`, `CREATED_AT`, `UPDATED_AT` FROM `brand_document` WHERE `ID` = :p0';
+        $sql = 'SELECT `ID`, `BRAND_ID`, `FILE`, `VISIBLE`, `POSITION`, `CREATED_AT`, `UPDATED_AT` FROM `brand_document` WHERE `ID` = :p0';
         try {
             $stmt = $con->prepare($sql);
             $stmt->bindValue(':p0', $key, PDO::PARAM_INT);
@@ -352,6 +356,47 @@ abstract class BrandDocumentQuery extends ModelCriteria
         }
 
         return $this->addUsingAlias(BrandDocumentTableMap::FILE, $file, $comparison);
+    }
+
+    /**
+     * Filter the query on the visible column
+     *
+     * Example usage:
+     * <code>
+     * $query->filterByVisible(1234); // WHERE visible = 1234
+     * $query->filterByVisible(array(12, 34)); // WHERE visible IN (12, 34)
+     * $query->filterByVisible(array('min' => 12)); // WHERE visible > 12
+     * </code>
+     *
+     * @param     mixed $visible The value to use as filter.
+     *              Use scalar values for equality.
+     *              Use array values for in_array() equivalent.
+     *              Use associative array('min' => $minValue, 'max' => $maxValue) for intervals.
+     * @param     string $comparison Operator to use for the column comparison, defaults to Criteria::EQUAL
+     *
+     * @return ChildBrandDocumentQuery The current query, for fluid interface
+     */
+    public function filterByVisible($visible = null, $comparison = null)
+    {
+        if (is_array($visible)) {
+            $useMinMax = false;
+            if (isset($visible['min'])) {
+                $this->addUsingAlias(BrandDocumentTableMap::VISIBLE, $visible['min'], Criteria::GREATER_EQUAL);
+                $useMinMax = true;
+            }
+            if (isset($visible['max'])) {
+                $this->addUsingAlias(BrandDocumentTableMap::VISIBLE, $visible['max'], Criteria::LESS_EQUAL);
+                $useMinMax = true;
+            }
+            if ($useMinMax) {
+                return $this;
+            }
+            if (null === $comparison) {
+                $comparison = Criteria::IN;
+            }
+        }
+
+        return $this->addUsingAlias(BrandDocumentTableMap::VISIBLE, $visible, $comparison);
     }
 
     /**