+++ core/lib/Thelia/Model/Base/FolderImageQuery.php
@@ -25,6 +25,7 @@ use Thelia\Model\Map\FolderImageTableMap;
  * @method     ChildFolderImageQuery orderById($order = Criteria::ASC) Order by the id column
  * @method     ChildFolderImageQuery orderByFolderId($order = Criteria::ASC) Order by the folder_id column
  * @method     ChildFolderImageQuery orderByFile($order = Criteria::ASC) Order by the file column
+ * @method     ChildFolderImageQuery orderByVisible($order = Criteria::ASC) Order by the visible column
  * @method     ChildFolderImageQuery orderByPosition($order = Criteria::ASC) Order by the position column
  * @method     ChildFolderImageQuery orderByCreatedAt($order = Criteria::ASC) Order by the created_at column
  * @method     ChildFolderImageQuery orderByUpdatedAt($order = Criteria::ASC) Order by the updated_at column
@@ -32,6 +33,7 @@ use Thelia\Model\Map\FolderImageTableMap;
  * @method     ChildFolderImageQuery groupById() Group by the id column
  * @method     ChildFolderImageQuery groupByFolderId() Group by the folder_id column
  * @method     ChildFolderImageQuery groupByFile() Group by the file column
+ * @method     ChildFolderImageQuery groupByVisible() Group by the visible column
  * @method     ChildFolderImageQuery groupByPosition() Group by the position column
  * @method     ChildFolderImageQuery groupByCreatedAt() Group by the created_at column
  * @method     ChildFolderImageQuery groupByUpdatedAt() Group by the updated_at column
@@ -54,6 +56,7 @@ use Thelia\Model\Map\FolderImageTableMap;
  * @method     ChildFolderImage findOneById(int $id) Return the first ChildFolderImage filtered by the id column
  * @method     ChildFolderImage findOneByFolderId(int $folder_id) Return the first ChildFolderImage filtered by the folder_id column
  * @method     ChildFolderImage findOneByFile(string $file) Return the first ChildFolderImage filtered by the file column
+ * @method     ChildFolderImage findOneByVisible(int $visible) Return the first ChildFolderImage filtered by the visible column
  * @method     ChildFolderImage findOneByPosition(int $position) Return the first ChildFolderImage filtered by the position column
  * @method     ChildFolderImage findOneByCreatedAt(string $created_at) Return the first ChildFolderImage filtered by the created_at column
  * @method     ChildFolderImage findOneByUpdatedAt(string $updated_at) Return the first ChildFolderImage filtered by the updated_at column
@@ -61,6 +64,7 @@ use Thelia\Model\Map\FolderImageTableMap;
  * @method     array findById(int $id) Return ChildFolderImage objects filtered by the id column
  * @method     array findByFolderId(int $folder_id) Return ChildFolderImage objects filtered by the folder_id column
  * @method     array findByFile(string $file) Return ChildFolderImage objects filtered by the file column
+ * @method     array findByVisible(int $visible) Return ChildFolderImage objects filtered by the visible column
  * @method     array findByPosition(int $position) Return ChildFolderImage objects filtered by the position column
  * @method     array findByCreatedAt(string $created_at) Return ChildFolderImage objects filtered by the created_at column
  * @method     array findByUpdatedAt(string $updated_at) Return ChildFolderImage objects filtered by the updated_at column
@@ -152,7 +156,7 @@ abstract class FolderImageQuery extends ModelCriteria
      */
     protected function findPkSimple($key, $con)
     {
-        $sql = 'SELECT `ID`, `FOLDER_ID`, `FILE`, `POSITION`, `CREATED_AT`, `UPDATED_AT` FROM `folder_image` WHERE `ID` = :p0';
+        $sql = 'SELECT `ID`, `FOLDER_ID`, `FILE`, `VISIBLE`, `POSITION`, `CREATED_AT`, `UPDATED_AT` FROM `folder_image` WHERE `ID` = :p0';
         try {
             $stmt = $con->prepare($sql);
             $stmt->bindValue(':p0', $key, PDO::PARAM_INT);
@@ -352,6 +356,47 @@ abstract class FolderImageQuery extends ModelCriteria
         }
 
         return $this->addUsingAlias(FolderImageTableMap::FILE, $file, $comparison);
+    }
+
+    /**
+     * Filter the query on the visible column
+     *
+     * Example usage:
+     * <code>
+     * $query->filterByVisible(1234); // WHERE visible = 1234
+     * $query->filterByVisible(array(12, 34)); // WHERE visible IN (12, 34)
+     * $query->filterByVisible(array('min' => 12)); // WHERE visible > 12
+     * </code>
+     *
+     * @param     mixed $visible The value to use as filter.
+     *              Use scalar values for equality.
+     *              Use array values for in_array() equivalent.
+     *              Use associative array('min' => $minValue, 'max' => $maxValue) for intervals.
+     * @param     string $comparison Operator to use for the column comparison, defaults to Criteria::EQUAL
+     *
+     * @return ChildFolderImageQuery The current query, for fluid interface
+     */
+    public function filterByVisible($visible = null, $comparison = null)
+    {
+        if (is_array($visible)) {
+            $useMinMax = false;
+            if (isset($visible['min'])) {
+                $this->addUsingAlias(FolderImageTableMap::VISIBLE, $visible['min'], Criteria::GREATER_EQUAL);
+                $useMinMax = true;
+            }
+            if (isset($visible['max'])) {
+                $this->addUsingAlias(FolderImageTableMap::VISIBLE, $visible['max'], Criteria::LESS_EQUAL);
+                $useMinMax = true;
+            }
+            if ($useMinMax) {
+                return $this;
+            }
+            if (null === $comparison) {
+                $comparison = Criteria::IN;
+            }
+        }
+
+        return $this->addUsingAlias(FolderImageTableMap::VISIBLE, $visible, $comparison);
     }
 
     /**