+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 #2.1.0-alpha1
 - Added sale management function
-- - Added `module_id` parameter to Area loop
+- Added `module_id` parameter to Area loop
 - Added "Shipping configuration" button to the delivery module list, with a warning if no shipping zone is assigned to the module.
 - Added the `show_label` parameter to the `render_form_field Smarty` function.
+- Added visible argument to image/document classes
 
 ##DEPRECATED
 Redirect methods are deprecated. You have now two way for redirecting a user :